(again, potentially) fixed theme hotswitching in gh-pages
Updated URL to use window.location.origin, as opposed to a hardcoded path. This would likely already work if the pages instance is hosted on the user's repo (x.github.io/), as opposed to in a repo (x.github.io/bento-next), as it's looking for the theme at the url base, not the hosted url.